+++ src/Pim/Bundle/DataGridBundle/Datasource/ResultRecord/MongoDbOdm/ObjectHydrator.php
@@ -5,23 +5,21 @@ namespace Pim\Bundle\DataGridBundle\Datasource\ResultRecord\MongoDbOdm;
 use Pim\Bundle\DataGridBundle\Datasource\ResultRecord\HydratorInterface;
 
 /**
- * Hydrate results of Doctrine ODM query as array of ids
+ * Object hydrator for ODM implementation
  *
  * @author    Romain Monceau <romain@akeneo.com>
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class ObjectIdHydrator implements HydratorInterface
+class ObjectHydrator implements HydratorInterface
 {
     /**
      * {@inheritdoc}
      */
     public function hydrate($qb, $options)
     {
-        $qb->hydrate(false)->select('_id');
+        $cursor = $qb->getQuery()->execute();
 
-        $results = $qb->getQuery()->execute();
-
-        return array_keys($results->toArray());
+        return $cursor->toArray();
     }
 }